requirements.txt
flask
Flask-Session
requests
openai
langchain
langchain-openai
PyPDF2
python-dotenv
pillow
cassio
BeautifulSoup4
python-docx